Refactored App.py
