+++ backend-flask/app.py
@@ -1,8 +1,17 @@
+import os
+import sys
+
 from flask import Flask
 from flask import request,  g
 from flask_cors import CORS, cross_origin
-import os
-import sys
+
+
+from lib.rollbar import init_rollbar
+from lib.xray import init_xray
+from lib.honeycomb import init_honeycomb
+from lib.cors import init_cors
+from lib.cloudwatch import init_cloudwatch
+from lib.cognito_jwt_token import jwt_required
 
 from services.users_short import *
 from services.home_activities import *
@@ -20,61 +29,17 @@ from services.update_profile import *
 
 #from lib.cognito_token_verification import CognitoTokenVerification
 
-from lib.cognito_jwt_token import jwt_required
 
 
-# HoneyComb ---------
-from opentelemetry import trace
-from opentelemetry.instrumentation.flask import FlaskInstrumentor
-from opentelemetry.instrumentation.requests import RequestsInstrumentor
-from opentelemetry.exporter.otlp.proto.http.trace_exporter import OTLPSpanExporter
-from opentelemetry.sdk.trace import TracerProvider
-from opentelemetry.sdk.trace.export import BatchSpanProcessor
-from opentelemetry.sdk.trace.export import ConsoleSpanExporter, SimpleSpanProcessor
-
-# X-RAY ----------
-from aws_xray_sdk.core import xray_recorder
-from aws_xray_sdk.ext.flask.middleware import XRayMiddleware
-
-# CloudWatch ------>
-import watchtower
-import logging
 
 
-# Rollbar import ------>
-from time import strftime
-
-import os
-import rollbar
-import rollbar.contrib.flask
-from flask import got_request_exception
-
-# Configuring Logger to Use CloudWatch
-# LOGGER = logging.getLogger(__name__)
-# LOGGER.setLevel(logging.DEBUG)
-# console_handler = logging.StreamHandler()
-# cw_handler = watchtower.CloudWatchLogHandler(log_group='cruddur')
-# LOGGER.addHandler(console_handler)
-# LOGGER.addHandler(cw_handler)
-# LOGGER.info("test-logs")
 
 
-# HoneyComb ---------
-# Initialize tracing and an exporter that can send data to Honeycomb
-provider = TracerProvider()
-processor = BatchSpanProcessor(OTLPSpanExporter())
-provider.add_span_processor(processor)
 
-# X-RAY ----------
-xray_url = os.getenv("AWS_XRAY_URL")
-xray_recorder.configure(service='backend-flask', dynamic_naming=xray_url)
 
-# Show this in the logs within the backend-flask app (STDOUT)
-simple_processor = SimpleSpanProcessor(ConsoleSpanExporter())
-provider.add_span_processor(simple_processor)
 
-trace.set_tracer_provider(provider)
-tracer = trace.get_tracer(__name__)
+
+
 
 app = Flask(__name__)
 
@@ -85,32 +50,13 @@ app = Flask(__name__)
 # )
 
 
-    # X-RAY ----------
-XRayMiddleware(app, xray_recorder)
-
-# HoneyComb ---------
-# Initialize automatic instrumentation with Flask
-FlaskInstrumentor().instrument_app(app)
-RequestsInstrumentor().instrument()
-
-# Rollbar init code
-rollbar_access_token = os.getenv('ROLLBAR_ACCESS_TOKEN')
-#@app.before_first_request
+## Initialization ==========>
+init_xray()
 with app.app_context():
-    def init_rollbar():
-        """init rollbar module"""
-        rollbar.init(
-        # access token
-            rollbar_access_token,
-        # environment name
-        'production',
-        # server root directory, makes tracebacks prettier
-            root=os.path.dirname(os.path.realpath(__file__)),
-        # flask already sets up logging
-            allow_logging_basic_config=False)
+    rollbar = init_rollbar()
 
-        # send exceptions from `app` to rollbar, using flask's signal system.
-        got_request_exception.connect(rollbar.contrib.flask.report_exception, app)
+init_honeycomb(app)
+init_cors(app)
 
 
 
@@ -122,26 +68,11 @@ with app.app_context():
     #  methods="OPTIONS,GET,HEAD,POST"
     # )
 
-    # Cloudwatch ------>
-    # @app.after_request
-    # def after_request(response):
-    #     timestamp = strftime('[%Y-%b-%d %H:%M]')
-    #     LOGGER.error('%s %s %s %s %s %s', timestamp, request.remote_addr, request.method, request.scheme, request.full_path, response.status)
-    #     return response
+    
 
-    # Rollbar init code
 
-frontend = os.getenv('FRONTEND_URL')
-backend = os.getenv('BACKEND_URL')
-origins = [frontend, backend]
 
-cors = CORS(
-app,
-resources={r"/api/*": {"origins": origins}},
-headers=['Content-Type', 'Authorization'],
-expose_headers='Authorization',
-methods="OPTIONS,GET,HEAD,POST"
-)
+
 
 def model_json(model):
   if model['errors'] is not None:
@@ -165,10 +96,7 @@ def rollbar_test():
 @jwt_required()
 def data_message_groups():
     model = MessageGroups.run(cognito_user_id=g.cognito_user_id)
-    if model['errors'] is not None:
-        return model['errors'], 422
-    else:
-        return model['data'], 200
+    return model_json(model)
 
 
 
@@ -179,12 +107,8 @@ def data_messages(message_group_uuid):
         cognito_user_id=g.cognito_user_id,
         message_group_uuid=message_group_uuid
         )
-    if model['errors'] is not None:
-        return model['errors'], 422
-    else:
-        return model['data'], 200
-    
-    
+    return model_json(model)    
+        
 
 @app.route("/api/profile/update", methods=['POST','OPTIONS'])
 @cross_origin()
@@ -197,10 +121,7 @@ def data_update_profile():
       bio=bio,
       display_name=display_name
   )
-  if model['errors'] is not None:
-    return model['errors'], 422
-  else:
-    return model['data'], 200
+  return model_json(model)
   
 
 @app.route("/api/messages", methods=['POST', 'OPTIONS'])
@@ -226,11 +147,7 @@ def data_create_message():
          message_group_uuid=message_group_uuid,
          cognito_user_id=g.cognito_user_id
       )
-    if model['errors'] is not None:
-        return model['errors'], 422
-    else:
-        return model['data'], 200
-    
+      return model_json(model)    
 
 def default_home_feed(e):
   # unauthenicatied request
@@ -257,21 +174,14 @@ def data_notifications():
 @xray_recorder.capture('activities-for-user-handle')
 def data_handle(handle):
     model = UserActivities.run(handle)
-    if model['errors'] is not None:
-        return model['errors'], 422
-    else:
-        return model['data'], 200
+    return model_json(model)
 
 
 @app.route("/api/activities/search", methods=['GET'])
 def data_search():
     term = request.args.get('term')
     model = SearchActivities.run(term)
-    if model['errors'] is not None:
-        return model['errors'], 422
-    else:
-        return model['data'], 200
-    return
+    return model_json(model)
 
 @app.route("/api/activities", methods=['POST', 'OPTIONS'])
 @cross_origin()
@@ -280,11 +190,7 @@ def data_activities():
     message = request.json['message']
     ttl = request.json['ttl']
     model = CreateActivity.run(message, g.cognito_user_id, ttl)
-    if model['errors'] is not None:
-        return model['errors'], 422
-    else:
-        return model['data'], 200
-    
+    return model_json(model)
 
 
 @app.route("/api/activities/<string:activity_uuid>", methods=['GET'])
@@ -295,14 +201,10 @@ def data_show_activity(activity_uuid):
 @app.route("/api/activities/<string:activity_uuid>/reply", methods=['POST', 'OPTIONS'])
 @cross_origin()
 def data_activities_reply(activity_uuid):
-    user_handle = 'andrewbrown'
+    user_handle = 'michaelcorleone'
     message = request.json['message']
     model = CreateReply.run(message, user_handle, activity_uuid)
-    if model['errors'] is not None:
-        return model['errors'], 422
-    else:
-        return model['data'], 200
-    return
+    return model_json(model)
 
 @app.route("/api/users/@<string:handle>/short", methods=['GET'])
 def data_users_short(handle):
