+++ backend-flask/lib/cognito_jwt_token.py
@@ -9,7 +9,7 @@ from jose.utils import base64url_decode
 from functools import wraps, partial
 from flask import request, g
 import os
-import app
+from flask import current_app as app
 
 class FlaskAWSCognitoError(Exception):
     pass
@@ -143,7 +143,7 @@ def jwt_required(f=None, on_error=None):
             # unauthenticated request
             app.logger.debug(e)
             if on_error:
-                on_error(e)
+               return on_error(e)
             return {}, 401
         return f(*args, **kwargs)
     return decorated_function
